requirements.txt
beautifulsoup4==4.11.1
Flask==2.1.0
requests==2.23.0
gunicorn
python-dotenv
